Added aarch64 workflow
Given that the vast majority of our target devices use the aarch64 ISA,
it would be beneficial to run a build and test using that architecture.
Since Github runners use x64, we have to use qemu to emulate both the
build and test environment. There is a signifact performance penalty
when using qemu. It might be possible to do the build using the
aarch64-unknown-linux-gnu toolchain on x64 and just run the unit tests
in the qemu environment, but it would require the sysroot and rootfs to
be binary compatible.