Remove unwraps
Calling unwrap() is an anti-pattern in Rust (with limited exceptions).
It is better to use pattern matching or other provided functions to
handle the cases where the enum is not the expected variant. This also
has the added benefit of reducing code size since the panic!() macro to
handle the unexpected variant will not be generated.